pyproject.toml
[tool.poetry]
name = "spot-gab-api"
version = "0.1.0"
description = "spot-gab-api"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0.post1"
alembic = "^1.13.1" # マイグレーションツール
python-dotenv = "^1.0.1"
injector = "^0.21.0" #DIライブラリ
pydantic-i18n = "^0.4.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
sqlmodel = "^0.0.14" # ORM
sqlalchemy-utils = "^0.41.1"
aiomysql = "^0.2.0"
python-multipart = "^0.0.9"
fastapi-mail = "^1.4.1"
jinja2 = "^3.1.3"
aiofiles = "^23.2.1"
python-i18n = "^0.3.9"
geoalchemy2 = "^0.14.7"
websockets = "^12.0"


[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.5"
aiosqlite = "^0.19.0"
httpx = "^0.26.0"
pytest = "^8.0.0"
pytest-mock = "^3.12.0"
isort = "^5.13.2"
black = "^24.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 119

[tool.isort]
profile = "black"
line_length = 119